pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "rocketshp"
version = "0.0.1"
description = "Fast structural heterogeneity estimation"
authors = [
  { name = "Samuel Sledzieski" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
requires-python = "~=3.11"

dependencies = [
  "biopython>=1.79",
  "biotite>=0.41.2",
  "datasets",
  "esm>=3.1.3",
  "h5py",
  "huggingface_hub",
  "lightning>=2.4.0",
  "loguru",
  "matplotlib",
  "neptune>=1.13.0",
  "nglview",
  "numpy>=1.23.5",
  "numpy-indexed>=0.3.7",
  "omegaconf",
  "pandas",
  "python-dateutil",
  "python-dotenv",
  "scikit_learn",
  "scipy",
  "seaborn",
  "statsmodels",
  "tokenizers",
  "torchmetrics",
  "transformers",
  "tqdm",
  "typer",
]

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501", "C"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[project.scripts]
rocketshp_train = "rocketshp.interface.train:__app__"
rocketshp_predict = "rocketshp.interface.predict:__app__"